fix for bug refs #608

+++ appion/appionlib/apCtf.py
@@ -41,8 +41,14 @@ def commitCtfValueToDatabase(imgdict, matlab, ctfvalue, params):
 	opimfilepath1 = os.path.join(params['opimagedir'],opimfile1)
 	opimfilepath2 = os.path.join(params['opimagedir'],opimfile2)
 
-	shutil.copyfile(imfile1, opimfilepath1)
-	shutil.copyfile(imfile2, opimfilepath2)
+	if os.path.isfile(imfile1):
+		shutil.copyfile(imfile1, opimfilepath1)
+	else:
+		apDisplay.printWarning("imfile1 is missing, %s"%(imfile1))
+	if os.path.isfile(imfile2):
+		shutil.copyfile(imfile2, opimfilepath2)
+	else:
+		apDisplay.printWarning("imfile2 is missing, %s"%(imfile2))
 	#pymat.eval(matlab,"im1 = imread('"+imfile1+"');")
 	#pymat.eval(matlab,"im2 = imread('"+imfile2+"');")
 	#pymat.eval(matlab,"imwrite(im1,'"+opimfilepath1+"');")
